Add files via upload
Here we have used GridSearchCV on the Iris flower dataset. 
By hyperparameter tuning we have come to the connclusion that the svm model is the most prefect machine learning model for this particular dataset as it gives us a 98% accuracy! 
Gridsearch conducted using the following algorithms :
1)SVM
2)logistic regression
3)Decision Tree classifier
4)Random Forest Classifier

(I personally didn't inlcude Gaussian or Multinomial Naive bayes here. But you can try that one too!)